Distributed Multisig Credential Issuance via a Kiwi endpoint.
Credential registry inception now multisig aware.

Credential registry incept from the command line now works for multisig.

Credential issuance from the command line KIWI endpoint now works for multisig

Several additional demo commands to bootstrap 3 multisig agents with a shared credential registry.
