+++ src/keri/db/basing.py
@@ -130,6 +130,7 @@ class GroupIdRecord:  # baser.gids
     """
     lid: str  # local identifier that contributes to the group
     gid: str  # group identifier prefix
+    dig: str  # qb64 of latest digest in the group
     cst: str  # group signing threshold of the next key commitment
     aids: list  # all identifiers participating in the group identity
 
@@ -282,29 +283,6 @@ class LocationRecord:  # baser.locs
 
     def __iter__(self):
         return iter(asdict(self))
-
-
-@dataclass
-class WitnessRecord:  # wits
-    """
-    Tracks the last message index retrieved from the witness mailbox
-    Database Key is the identifier prefix of the witness that is storing
-    events in a mailbox
-    """
-    topics: dict
-
-
-@dataclass
-class GroupIdentifier:  # gids
-    """
-    Track group identifiers that we are participating in
-    Database Key is the identifier prefix of the group identifier
-    """
-    lid: str  # local identifier that contributes to the group
-    gid: str  # group identifier prefix
-    dig: str  # qb64 of latest digest in the group
-    cst: str  # group signing threshold of the next key commitment
-    aids: list  # all identifiers participating in the group identity
 
 
 def openDB(name="test", **kwa):
